戦闘終了時の処理を別ファイルに切り出した
